Fix potential race condition with on_command()
It may be possible for send_command() to not receive a timely response, which leads it to send a second packet, after which the protocol could receive a response for both packets.
This makes _dispatch_packet() check the command queue beforehand to guard against duplicates like send_command() does.
Accordingly, keep alive packets now add themselves to the queue.